+++ baybe/searchspace/continuous.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import gc
 import warnings
 from collections.abc import Collection, Sequence
-from itertools import chain
+from itertools import chain, repeat
 from typing import TYPE_CHECKING, Any, cast
 
 import numpy as np
@@ -16,6 +16,7 @@ from typing_extensions import override
 from baybe.constraints import (
     ContinuousCardinalityConstraint,
     ContinuousLinearConstraint,
+    ContinuousLinearInterPointConstraint,
 )
 from baybe.constraints.base import ContinuousConstraint, ContinuousNonlinearConstraint
 from baybe.constraints.validation import (
@@ -68,6 +69,16 @@ class SubspaceContinuous(SerialMixin):
     )
     """Nonlinear constraints."""
 
+    constraints_ip_lin_eq: tuple[ContinuousLinearInterPointConstraint, ...] = field(
+        converter=to_tuple, factory=tuple
+    )
+    "Linear interpoint equality constraints."
+
+    constraints_ip_lin_ineq: tuple[ContinuousLinearInterPointConstraint, ...] = field(
+        converter=to_tuple, factory=tuple
+    )
+    "Linear interpoint inequality constraints."
+
     @override
     def __str__(self) -> str:
         if self.is_empty:
@@ -82,15 +93,35 @@ class SubspaceContinuous(SerialMixin):
         nonlin_constraints_list = [
             constr.summary() for constr in self.constraints_nonlin
         ]
+        nonlin_constraints_list = [
+            constr.summary() for constr in self.constraints_nonlin
+        ]
+        ip_eq_constraints_list = [
+            constr.summary() for constr in self.constraints_ip_lin_eq
+        ]
+        ip_ineq_constraints_list = [
+            constr.summary() for constr in self.constraints_ip_lin_ineq
+        ]
+
         param_df = pd.DataFrame(param_list)
         lin_eq_df = pd.DataFrame(eq_constraints_list)
         lin_ineq_df = pd.DataFrame(ineq_constraints_list)
+        ip_lin_eq_df = pd.DataFrame(ip_eq_constraints_list)
+        ip_lin_ineq_df = pd.DataFrame(ip_ineq_constraints_list)
         nonlinear_df = pd.DataFrame(nonlin_constraints_list)
 
         fields = [
             to_string("Continuous Parameters", pretty_print_df(param_df)),
             to_string("Linear Equality Constraints", pretty_print_df(lin_eq_df)),
             to_string("Linear Inequality Constraints", pretty_print_df(lin_ineq_df)),
+            to_string(
+                "Linear Interpoint Equality Constraints",
+                pretty_print_df(ip_lin_eq_df),
+            ),
+            to_string(
+                "Linear Interpoint Inequality Constraints",
+                pretty_print_df(ip_lin_ineq_df),
+            ),
             to_string("Non-linear Constraints", pretty_print_df(nonlinear_df)),
         ]
 
@@ -127,6 +158,32 @@ class SubspaceContinuous(SerialMixin):
         if any(c.is_eq for c in lst):
             raise ValueError(
                 f"The list '{fields(self.__class__).constraints_lin_ineq.name}' of "
+                f"{self.__class__.__name__} only accepts inequality constraints, i.e. "
+                f"the 'operator' for all list items should be '>=' or '<='."
+            )
+
+    @constraints_ip_lin_eq.validator
+    def _validate_constraints_ip_lin_eq(
+        self, _, lst: list[ContinuousLinearInterPointConstraint]
+    ) -> None:
+        """Validate linear interpoint equality constraints."""
+        # TODO Remove once eq and ineq constraints are consolidated into one list
+        if not all(c.is_eq for c in lst):
+            raise ValueError(
+                f"The list '{fields(self.__class__).constraints_ip_lin_eq.name}' of "
+                f"{self.__class__.__name__} only accepts equality constraints, i.e. "
+                f"the 'operator' for all list items should be '='."
+            )
+
+    @constraints_ip_lin_ineq.validator
+    def _validate_constraints_ip_lin_ineq(
+        self, _, lst: list[ContinuousLinearConstraint]
+    ) -> None:
+        """Validate linear interpoint inequality constraints."""
+        # TODO Remove once eq and ineq constraints are consolidated into one list
+        if any(c.is_eq for c in lst):
+            raise ValueError(
+                f"The list '{fields(self.__class__).constraints_ip_lin_ineq.name}' of "
                 f"{self.__class__.__name__} only accepts inequality constraints, i.e. "
                 f"the 'operator' for all list items should be '>=' or '<='."
             )
@@ -175,12 +232,30 @@ class SubspaceContinuous(SerialMixin):
             constraints_lin_eq=[  # type:ignore[attr-misc]
                 c
                 for c in constraints
-                if (isinstance(c, ContinuousLinearConstraint) and c.is_eq)
+                if (
+                    isinstance(c, ContinuousLinearConstraint)
+                    and not isinstance(c, ContinuousLinearInterPointConstraint)
+                    and c.is_eq
+                )
             ],
             constraints_lin_ineq=[  # type:ignore[attr-misc]
                 c
                 for c in constraints
-                if (isinstance(c, ContinuousLinearConstraint) and not c.is_eq)
+                if (
+                    isinstance(c, ContinuousLinearConstraint)
+                    and not isinstance(c, ContinuousLinearInterPointConstraint)
+                    and not c.is_eq
+                )
+            ],
+            constraints_ip_lin_eq=[  # type:ignore[misc]
+                c
+                for c in constraints
+                if (isinstance(c, ContinuousLinearInterPointConstraint) and c.is_eq)
+            ],
+            constraints_ip_lin_ineq=[  # type:ignore[misc]
+                c
+                for c in constraints
+                if (isinstance(c, ContinuousLinearInterPointConstraint) and not c.is_eq)
             ],
             constraints_nonlin=[  # type:ignore[attr-misc]
                 c for c in constraints if isinstance(c, ContinuousNonlinearConstraint)
@@ -282,6 +357,24 @@ class SubspaceContinuous(SerialMixin):
             index=["min", "max"],
         )
 
+    @property
+    def is_constrained(self) -> bool:
+        """Return whether the subspace is constrained in any way."""
+        return any(
+            (
+                self.constraints_ip_lin_eq,
+                self.constraints_ip_lin_ineq,
+                self.constraints_lin_eq,
+                self.constraints_lin_ineq,
+                self.constraints_nonlin,
+            )
+        )
+
+    @property
+    def has_interpoint_constraints(self) -> bool:
+        """Return whether or not the space has any interpoint constraints."""
+        return any((self.constraints_ip_lin_eq, self.constraints_ip_lin_ineq))
+
     def _drop_parameters(self, parameter_names: Collection[str]) -> SubspaceContinuous:
         """Create a copy of the subspace with certain parameters removed.
 
@@ -359,7 +452,10 @@ class SubspaceContinuous(SerialMixin):
         )
         return self.sample_uniform(n_points)
 
-    def sample_uniform(self, batch_size: int = 1) -> pd.DataFrame:
+    def sample_uniform(
+        self,
+        batch_size: int = 1,
+    ) -> pd.DataFrame:
         """Draw uniform random parameter configurations from the continuous space.
 
         Args:
@@ -383,14 +479,17 @@ class SubspaceContinuous(SerialMixin):
 
         if not self.parameters:
             return pd.DataFrame(index=pd.RangeIndex(0, batch_size))
-
-        if (
-            len(self.constraints_lin_eq) == 0
-            and len(self.constraints_lin_ineq) == 0
-            and len(self.constraints_cardinality) == 0
-        ):
+        # If the space is completely unconstrained, we can sample from bounds.
+        if not self.is_constrained:
             return self._sample_from_bounds(batch_size, self.comp_rep_bounds.values)
 
+        if self.has_interpoint_constraints:
+            return self._sample_from_polytope_with_interpoint_constraints(
+                batch_size, self.comp_rep_bounds.values
+            )
+
+        # If there are neither cardinality nor interpoint constraints, we sample
+        # directly from the polytope
         if len(self.constraints_cardinality) == 0:
             return self._sample_from_polytope(batch_size, self.comp_rep_bounds.values)
 
@@ -402,6 +501,77 @@ class SubspaceContinuous(SerialMixin):
             low=bounds[0, :], high=bounds[1, :], size=(batch_size, len(self.parameters))
         )
 
+        return pd.DataFrame(points, columns=self.parameter_names)
+
+    def _sample_from_polytope_with_interpoint_constraints(
+        self,
+        batch_size: int,
+        bounds: np.ndarray,
+    ) -> pd.DataFrame:
+        """Draw uniform random samples from a polytope with interpoint constraints."""
+        # If the space has interpoint constraints, we need to sample from a larger
+        # searchspace that models the batch size via additional dimension. This is
+        # necessary since `get_polytope_samples` cannot handle inter-point constraints,
+        # see https://github.com/pytorch/botorch/issues/2468
+
+        import torch
+        from botorch.utils.sampling import get_polytope_samples
+
+        from baybe.utils.numerical import DTypeFloatNumpy
+        from baybe.utils.torch import DTypeFloatTorch
+
+        # The number of parameters is needed at some places for adjusting indices
+        num_of_params = len(self.parameters)
+
+        eq_constraints, ineq_constraints = [], []
+
+        # We start with the general constraints before going to interpoint constraints
+        for c in [*self.constraints_lin_eq, *self.constraints_lin_ineq]:
+            param_indices, coefficients, rhs = c.to_botorch(self.parameters)
+            for b in range(batch_size):
+                botorch_tuple = (param_indices + b * num_of_params, coefficients, rhs)
+                if c.is_eq:
+                    eq_constraints.append(botorch_tuple)
+                else:
+                    ineq_constraints.append(botorch_tuple)
+
+        if self.has_interpoint_constraints:
+            for c in [
+                *self.constraints_ip_lin_eq,
+                *self.constraints_ip_lin_ineq,
+            ]:
+                # Get the indices of the parameters used in the constraint
+                param_index = {
+                    name: self.parameter_names.index(name) for name in c.parameters
+                }
+                param_indices_list = [
+                    batch * num_of_params + param_index[param]
+                    for param in c.parameters
+                    for batch in range(batch_size)
+                ]
+                coefficients_list = list(
+                    chain(*zip(*repeat(c.coefficients, batch_size)))
+                )
+                botorch_tuple = (
+                    torch.tensor(param_indices_list),
+                    torch.tensor(coefficients_list, dtype=DTypeFloatTorch),
+                    np.asarray(c.rhs, dtype=DTypeFloatNumpy).item(),
+                )
+                if c.is_eq:
+                    eq_constraints.append(botorch_tuple)
+                else:
+                    ineq_constraints.append(botorch_tuple)
+
+        bounds_joint = torch.cat(
+            [torch.from_numpy(bounds) for _ in range(batch_size)], dim=-1
+        )
+        points = get_polytope_samples(
+            n=1,
+            bounds=bounds_joint,
+            equality_constraints=eq_constraints,
+            inequality_constraints=ineq_constraints,
+        )
+        points = points.reshape(batch_size, points.shape[-1] // batch_size)
         return pd.DataFrame(points, columns=self.parameter_names)
 
     def _sample_from_polytope(